requirements.txt
# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Browser Automation
selenium==4.15.2
webdriver-manager==4.0.1

# AI/ML
openai==1.3.5

# Environment Management
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Date/Time Parsing
python-dateutil==2.8.2

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Logging and Monitoring
structlog==23.2.0

# Development Tools (optional)
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Security
cryptography==41.0.7

# Additional utilities
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2